Refactor handlers out of main script
